New Currency logic & filters in migration.
