+++ includes/admin/downloads/dashboard-columns.php
@@ -87,7 +87,7 @@ function edd_render_download_columns( $column_name, $post_id ) {
 		case 'earnings':
 			if ( current_user_can( 'view_product_stats', $post_id ) ) {
 				echo '<a href="' . esc_url( admin_url( 'edit.php?post_type=download&page=edd-reports&view=downloads&download-id=' . $post_id ) ) . '">';
-					echo edd_currency_filter( edd_get_download_earnings_stats( $post_id ), edd_get_currency() );
+					echo edd_currency_filter( edd_format_amount( edd_get_download_earnings_stats( $post_id ) ) );
 				echo '</a>';
 			} else {
 				echo '-';
