+++ includes/admin/upgrades/v3/class-data-migrator.php
@@ -405,6 +405,20 @@ class Data_Migrator {
 			$user_info = array();
 		}
 
+		/**
+		 * Last change to filter payment meta before we use it!
+		 * Note: If modifying `cart_details`, then it's recommended that you first run
+		 * `EDD\Admin\Upgrades\v3\Data_Migrator::fix_possible_serialization()`
+		 * before making adjustments.
+		 *
+		 * @since 3.0
+		 *
+		 * @param array $payment_meta Payment meta.
+		 * @param int   $payment_id   ID of the payment.
+		 * @param array $meta         All post meta.
+		 */
+		$payment_meta = apply_filters( 'edd_30_migration_payment_meta', $payment_meta, $data->ID, $meta );
+
 		$order_number   = isset( $meta['_edd_payment_number'][0] ) ? $meta['_edd_payment_number'][0] : '';
 		$user_id        = isset( $meta['_edd_payment_user_id'][0] ) && ! empty( $meta['_edd_payment_user_id'][0] ) ? $meta['_edd_payment_user_id'][0] : 0;
 		$ip             = isset( $meta['_edd_payment_user_ip'][0] ) ? $meta['_edd_payment_user_ip'][0] : '';
@@ -646,6 +660,18 @@ class Data_Migrator {
 			'total'          => $total,
 		);
 
+		/**
+		 * Filters the data used to create the order.
+		 *
+		 * @since 3.0
+		 *
+		 * @param array $order_data   Order creation arguments.
+		 * @param array $payment_meta Payment meta.
+		 * @param array $cart_details Cart details.
+		 * @param array $meta         All payment meta.
+		 */
+		$order_data = apply_filters( 'edd_30_migration_order_creation_data', $order_data, $payment_meta, $cart_details, $meta );
+
 		$order_id = edd_add_order( $order_data );
 
 		// Save an un-matched tax rate in order meta.
@@ -856,6 +882,18 @@ class Data_Migrator {
 					'date_modified' => $data->post_modified_gmt,
 				);
 
+				/**
+				 * Filters the arguments used to create the order item.
+				 *
+				 * @since 1.0
+				 *
+				 * @param array $order_item_args Order item arguments.
+				 * @param array $cart_item       Original cart item.
+				 * @param array $payment_meta    Payment meta.
+				 * @param array $meta            All meta.
+				 */
+				$order_item_args = apply_filters( 'edd_30_migration_order_item_creation_data', $order_item_args, $cart_item, $payment_meta, $meta );
+
 				$order_item_id = edd_add_order_item( $order_item_args );
 
 				if ( ! empty( $cart_item['item_number']['options'] ) ) {
@@ -944,6 +982,19 @@ class Data_Migrator {
 							'tax'         => $tax,
 							'total'       => floatval( $fee['amount'] ) + $tax,
 						);
+
+						/**
+						 * Filters the arguments used to create an order item adjustment.
+						 *
+						 * @since 3.0
+						 *
+						 * @param array $adjustment_args Adjustment arguments for a fee.
+						 * @param array $fee             Original fee data.
+						 * @param array $cart_item       Cart item this fee is part of.
+						 * @param array $payment_meta    Payment meta.
+						 * @param array $meta            All meta.
+						 */
+						$adjustment_args = apply_filters( 'edd_30_migration_order_item_adjustment_creation_data', $adjustment_args, $fee, $cart_item, $payment_meta, $meta );
 
 						$adjustment_id = edd_add_order_adjustment( $adjustment_args );
 
@@ -1037,6 +1088,18 @@ class Data_Migrator {
 					'date_modified' => $data->post_modified_gmt,
 				);
 
+				/**
+				 * Filters the order adjustment arguments.
+				 *
+				 * @since 3.0
+				 *
+				 * @param array $adjustment_args Arguments used to create the order adjustment.
+				 * @param array $fee             Fee data.
+				 * @param array $payment_meta    Payment meta.
+				 * @param array $meta            All meta.
+				 */
+				$adjustment_args = apply_filters( 'edd_30_migration_order_adjustment_creation_data', $adjustment_args, $fee, $payment_meta, $meta );
+
 				$adjustment_id = edd_add_order_adjustment( $adjustment_args );
 
 				if ( ! empty( $refund_id ) ) {
@@ -1076,19 +1139,33 @@ class Data_Migrator {
 					continue;
 				}
 
-				edd_add_order_adjustment(
-					array(
-						'object_id'     => $order_id,
-						'object_type'   => 'order',
-						'type_id'       => $discount->id,
-						'type'          => 'discount',
-						'description'   => $discount->code,
-						'subtotal'      => $subtotal - $discount->get_discounted_amount( $subtotal ),
-						'total'         => $subtotal - $discount->get_discounted_amount( $subtotal ),
-						'date_created'  => $date_created_gmt,
-						'date_modified' => $data->post_modified_gmt,
-					)
+				$discount_args = array(
+					'object_id'     => $order_id,
+					'object_type'   => 'order',
+					'type_id'       => $discount->id,
+					'type'          => 'discount',
+					'description'   => $discount->code,
+					'subtotal'      => $subtotal - $discount->get_discounted_amount( $subtotal ),
+					'total'         => $subtotal - $discount->get_discounted_amount( $subtotal ),
+					'date_created'  => $date_created_gmt,
+					'date_modified' => $data->post_modified_gmt,
 				);
+
+				/**
+				 * Filters the arguments used to create a discount adjustment.
+				 *
+				 * @since 3.0
+				 *
+				 * @param array         $discount_args Order adjustment arguments.
+				 * @param \EDD_Discount $discount      Discount object.
+				 * @param float         $subtotal      Order subtotal.
+				 * @param array         $user_info     User info array.
+				 * @param array         $payment_meta  Payment meta.
+				 * @param array         $meta          All post meta.
+				 */
+				$discount_args = apply_filters( 'edd_30_migration_order_discount_creation_data', $discount_args, $discount, $subtotal, $user_info, $payment_meta, $meta );
+
+				edd_add_order_adjustment( $discount_args );
 			}
 		}
 
@@ -1197,7 +1274,7 @@ class Data_Migrator {
 	 *
 	 * @return mixed
 	 */
-	private static function fix_possible_serialization( $data ) {
+	public static function fix_possible_serialization( $data ) {
 		if ( ! is_array( $data ) && is_string( $data ) ) {
 			$data = substr_replace( $data, 'a', 0, 1 );
 		}
